pyproject.toml
[project]
name = "advspeech"
version = "0.1.0"
description = "adversarial attack on 0-shot voice cloning"
authors = [
    {name = "Renyi Yang"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.9"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
ipython = "*"
librosa = "*"
matplotlib = "*"
numpy = "*"
pandas = "*"
parselmouth = "*"
pesq = "*"
pyworld = "*"
PyYAML = "*"
tqdm = "*"

[tool.poetry]
virtualenvs.create = false
